fix(055): address PR #28 second review — critical EF FK + transaction fixes
Critical fixes:
- Remove polymorphic FK constraint from PasskeyCredentials.OwnerId → PublicIdentities.Id
  that would break org-user passkey registration (OwnerId is polymorphic, not a real FK)
- Fix misleading ID token comment — clarify TLS transport trust boundary

High-priority fixes:
- Fix transaction integrity gap in public user registration: VerifyRegistrationAsync
  now accepts persist param, public signup saves identity + credential in single transaction
- Squash EF migration into regenerated InitialCreate (not yet production)

Medium fixes:
- SignatureCounter: int → long to avoid lossy cast from FIDO2 uint SignCount
- Fido2 config startup validation (throw on missing ServerDomain/Origins)
- Extract GitHub email URL to constant (GitHubUserEmailsEndpoint)

+++ src/Services/Sorcha.Tenant.Service/Models/PasskeyCredential.cs
@@ -52,8 +52,9 @@ public class PasskeyCredential
     /// <summary>
     /// Signature counter for cloned authenticator detection.
     /// Increments on each authentication; regression indicates a cloned authenticator.
+    /// Stored as long to avoid lossy cast from FIDO2's uint SignCount.
     /// </summary>
-    public int SignatureCounter { get; set; } = 0;
+    public long SignatureCounter { get; set; } = 0;
 
     /// <summary>
     /// Type of the credential owner: "OrgUser" or "PublicIdentity".
